+++ scripts/lichessTools.js
@@ -172,6 +172,11 @@
       WhiteStar: '\u2606',
       BulletPoint: '\u2022',
       Ellipsis: '\u2026',
+      ThumbsUpSign: '\uD83D\uDC4D',
+      ThumbsDownSign: '\uD83D\uDC4E',
+      PoutingFace: '\uD83D\uDE21',
+      SmilingFaceWithHorns: '\uD83D\uDE08',
+      SparklingHeart: '\uD83D\uDC96',
 
       toEntity: function(s) {
         let result='';
@@ -423,6 +428,15 @@
         }
       }
     }
+
+    getPuzzleId() {
+      const $ = this.$;
+      const puzzleUrl = $('main.puzzle div.infos.puzzle a[href^="/training"]').attr('href');
+      if (!puzzleUrl) return;
+      const m = /\/training\/([^\/]+)$/.exec(puzzleUrl);
+      if (!m) return;
+      return m[1];
+    };
 
     unbindKeyHandler(combo, onlyMine) {
       const mousetrap = this.lichess.mousetrap;
@@ -1990,6 +2004,16 @@
           }
           return result;
         }
+      },
+      tournament: {
+        lichessTools: this,
+        getInfo: async function (tourId) {
+          const lt = this.lichessTools;
+          const userId = lt.getUserId();
+          if (!userId) return null;
+          const data = await lt.net.json({ url: '/tournament/{tourId}?page=1&partial=true&me={userId}', args: { tourId: tourId, userId:userId } });
+          return data;
+        }
       }
     }
 
@@ -2086,6 +2110,46 @@
       return options;
     }
 
+    newerVersion(v1,v2) {
+      if (!v1) return false;
+      if (!v2) return true;
+      const arr1 = v1.split('.');
+      const arr2 = v2.split('.');
+      const l = Math.max(arr1.length, arr2.length);
+      for (let i=0; i<l; i++) {
+        if ((arr1[i]||0)>(arr2[i]||0)) return true;
+      }
+      return false;
+    }
+
+    async upgradeOptions(options) {
+      for (const tool of this.tools) {
+        if (!tool.upgrades?.length) continue;
+        const upgrades = tool.upgrades.filter(u=>this.newerVersion(u.version,options.version));
+        for (const upgrade of upgrades) {
+          const pref = tool.preferences.find(p=>p.name == upgrade.name);
+          if (!pref) {
+            this.global.console.warn('Cannot find preference with name',upgrade.name);
+            continue;
+          }
+          if (pref.type != 'multiple') {
+            this.global.console.warn('Preference with name',upgrade.name,'is not multiple');
+            continue;
+          }
+          if (!this.isOptionSet(pref.defaultValue,upgrade.value)) {
+            this.global.console.warn('Preference value',upgrade.name+'.'+upgrade.value,'is not enabled by default');
+            continue;
+          }
+          let currentValue = options[upgrade.name];
+          if (!this.isOptionSet(currentValue,upgrade.value)) {
+            options[upgrade.name] = currentValue
+                                ? currentValue+','+upgrade.value
+                                : upgrade.value;
+          }
+        }
+      }
+    }
+
     async getOptions() {
       let options = this.global.localStorage.getItem('LiChessTools2.options');
       options = this.jsonParse(options);
@@ -2095,6 +2159,7 @@
         ...defaults,
         ...options
       };
+      await this.upgradeOptions(options);
       options.getValue = function (optionName, optionValue) {
         if (!this.enableLichessTools) return false;
         return this[optionName]
@@ -2131,6 +2196,11 @@
     }
 
     async saveOptions(options) {
+      const data = await this.comm.send({ type: 'getVersion' })
+                                             .catch(e => { throw e; });
+      const version = data.version;
+      this.debug && console.log('Saving options version',version);
+      options.version = version;
       const optionsJson = this.global.JSON.stringify(options);
       this.global.localStorage.setItem('LiChessTools2.options', optionsJson);
     }