+++ HISTORY.md
@@ -1,5 +1,24 @@
 History of features added to LiChess Tools in time (reversed order)
 
+26 Jan
+
+- chapter cls markers
+- tournament top board
+- emoji quick action
+
+22 Jan
+
+- started puzzle history
+
+21 Jan
+
+- adding more languages: Dutch, Arabic, Chinese Simplified
+
+16 Jan
+
+- option to export shapes in PGN
+- ability to upgrade older options with the new preferences
+
 15 Jan
 
 - show total game duration